requirements.txt
google-generativeai
python-dotenv
pytesseract
pillow
scipy
scikit-learn
scikit-image
imutils
matplotlib
requests
beautifulsoup4
h5py
tensorflow
textblob